[RDSS-1863] Enforce a stricter version field: SemVer 2.0.0
If tempted to simplify the regex, make sure that the following are forbidden

  1.2             (missing patch)
  1..             (empty version number)
  010.0.0         (leading zeroes)
  1.2.3-pre..     (empty components between dots of pre-release versions)
  1.2.3+build..   (empty components between dots in build metadata

Note that the following that looks like build metadata before pre-release version

  1.2.3+acd-efg

but this _is_ allowed, because the build metadata allows hypens, so the above
is build "acd-efg".

Note that from
https://spacetelescope.github.io/understanding-json-schema/reference/regular_expressions.html
that the usual complete regex syntax is not supported. Specifically word
boundaries are missing.